requirements.txt
streamlit
pandas
numpy
scipy
PyYaml
seaborn
-e .